Create README.md
The problem was to mine bitcoin which can be any random string that, when “hashed” produces an output smaller than a given target value (in practice, the comparison values have leading 0’s). It is very difficult to mine a bitcoin as the number of zeros increase on a single computer.
Solution - 
Created a distributed application based on the Erlang actor model in which multiple actors distributed over various client nodes generated hashed strings. The values are compared and if a bitcoin is found, it is sent back to server. 
The server and clients each run (4*No.of Cores) processes to ensure that all the cores are utilised completely. The string generation is an iterative approach where we are generating a workload of 10,000,000 for each process on the machine. Each process will be provided a start number and workload. The process will then mine for the bitcoin between start number and the workload. When a process in the server/client completes a given workload, it send a requests to server for the new workload. if a bitcoin is found, it is sent back to server.